+++ src/lrmatrix.c
@@ -168,7 +168,7 @@ int main_lrmatrix(int argc, char* argv[argc])
 
 	struct iter_admm_conf mmconf;
 	memcpy(&mmconf, &iter_admm_defaults, sizeof(struct iter_admm_conf));
-	mmconf.maxiter = (unsigned int)maxiter;
+	mmconf.maxiter = maxiter;
 	mmconf.rho = rho;
 	mmconf.hogwild = hogwild;
 	mmconf.fast = fast;
